Quaternion struct mostly done - missing generic functions. Also need to test it out.
Once this struct is done and more thorough tests are written, should create the 'transform' struct and start working on systems (while adding in what is needed for each as we go - but they all need a transform).

Start with collision/rendering and a super basic outline for GameObjects. Or, may want to start hooking up SDL2 so we can get some visuals.